--- packages/react-dom/npm/unstable-flight-server.js
@@ -1,3 +1,0 @@
-'use strict';
-
-module.exports = require('./unstable-flight-server.node');
