+++ packages/react-dom/package.json
@@ -39,18 +39,13 @@
     "unstable-fizz.js",
     "unstable-fizz.browser.js",
     "unstable-fizz.node.js",
-    "unstable-flight-client.js",
-    "unstable-flight-server.js",
-    "unstable-flight-server.browser.js",
-    "unstable-flight-server.node.js",
     "unstable-native-dependencies.js",
     "cjs/",
     "umd/"
   ],
   "browser": {
     "./server.js": "./server.browser.js",
-    "./unstable-fizz.js": "./unstable-fizz.browser.js",
-    "./unstable-flight-server.js": "./unstable-flight-server.browser.js"
+    "./unstable-fizz.js": "./unstable-fizz.browser.js"
   },
   "browserify": {
     "transform": [
