+++ packages/react-flight-dom-webpack/README.md
@@ -1,0 +1,5 @@
+# react-flight-dom-webpack
+
+Experimental React Flight bindings for DOM using Webpack.
+
+**Use it at your own risk.**
