+++ packages/react-flight-dom-webpack/npm/server.browser.js
@@ -1,0 +1,7 @@
+'use strict';
+
+if (process.env.NODE_ENV === 'production') {
+  module.exports = require('./cjs/react-flight-dom-webpack-server.browser.production.min.js');
+} else {
+  module.exports = require('./cjs/react-flight-dom-webpack-server.browser.development.js');
+}
