+++ packages/react-flight-dom-webpack/server.browser.js
@@ -9,8 +9,9 @@
 
 'use strict';
 
-const ReactFlightDOMServerNode = require('./src/server/flight/ReactFlightDOMServerNode');
+const ReactFlightDOMServerBrowser = require('./src/ReactFlightDOMServerBrowser');
 
 // TODO: decide on the top-level export form.
 // This is hacky but makes it work with both Rollup and Jest
-module.exports = ReactFlightDOMServerNode.default || ReactFlightDOMServerNode;
+module.exports =
+  ReactFlightDOMServerBrowser.default || ReactFlightDOMServerBrowser;
