+++ packages/react-flight-dom-webpack/server.node.js
@@ -9,9 +9,8 @@
 
 'use strict';
 
-const ReactFlightDOMServerBrowser = require('./src/server/flight/ReactFlightDOMServerBrowser');
+const ReactFlightDOMServerNode = require('./src/ReactFlightDOMServerNode');
 
 // TODO: decide on the top-level export form.
 // This is hacky but makes it work with both Rollup and Jest
-module.exports =
-  ReactFlightDOMServerBrowser.default || ReactFlightDOMServerBrowser;
+module.exports = ReactFlightDOMServerNode.default || ReactFlightDOMServerNode;
