+++ packages/react-flight/src/ReactFlightHostConfigBrowser.js
